+++ src/Dualis/Registries.cs
@@ -1,6 +1,7 @@
-using Dualis.CQRS;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
+
+using Dualis.CQRS;
 
 namespace Dualis;