+++ tests/SourceGen.Tests/CrossAssemblyDiscoveryTests.cs
@@ -1,6 +1,7 @@
 using System.Collections.Immutable;
 using System.Text;
 using Dualis.CQRS;
+using Dualis.SourceGen;
 using FluentAssertions;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;